fix: readiness bitmap uint32_t → uint64_t (was UB at 32+ clients)
ready_bitmap and connected_bitmap were uint32_t, causing undefined
behavior when client_idx >= 32 via (1u << client_idx). Despite
FACTORY_MAX_SIGNERS being 64, the readiness tracker silently broke
at 32+ clients. Changed to uint64_t with 1ULL shifts throughout.
Also updated popcount helper to operate on uint64_t.

+++ include/superscalar/readiness.h
@@ -20,8 +20,8 @@ typedef struct {
     readiness_entry_t clients[FACTORY_MAX_SIGNERS];
     size_t   n_clients;
     uint32_t factory_id;
-    uint32_t ready_bitmap;      /* bit i = connected AND ready */
-    uint32_t connected_bitmap;  /* bit i = connected */
+    uint64_t ready_bitmap;      /* bit i = connected AND ready */
+    uint64_t connected_bitmap;  /* bit i = connected */
     persist_t *db;              /* may be NULL */
 } readiness_tracker_t;
